test: split test_v016.py by feature; add real-file concurrency test (#98)
The version-named regression dump's 9 classes now live with their
features: PATH_PATTERN tests in test_watcher.py, disk-layer tests
(_infer_account_mailbox, link extraction, get_email_links) in
test_disk.py, MCP tool tests in test_server.py, _run_serve startup
in test_cli_profile.py. Test bodies unchanged.

Also adds the suite's first real-file SQLite contention test — the
shared fixtures use :memory:, which never locks, so busy_timeout
behavior under two concurrent writers was previously untested.

Closes #98
